+++ CHANGELOG.md
@@ -32,3 +32,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Blaze Store Exporter
 - Python to Docker
 - Blaze Id
+- .dockerignore
+- Github build.yml
+- Fetch Bundle from Blaze Store
+
+### Changed
+- Create temporal root directory if it does not exist